Geocoding

+++ db/migrate/20180222113013_add_columns_to_users.rb
@@ -1,6 +1,6 @@
 class AddColumnsToUsers < ActiveRecord::Migration[5.1]
   def change
-    add_column :users, :latitude, :integer
-    add_column :users, :longitude, :integer
+    add_column :users, :latitude, :float
+    add_column :users, :longitude, :float
   end
 end
